eh: use usleep_range instead of msleep
There are couple of issues with msleep

1. It's not accurate so easily observed it takes above 10ms.

Quote from [1]
```
SLEEPING FOR ~USECS OR SMALL MSECS ( 10us - 20ms):
Use usleep_range

Why not msleep for (1ms - 20ms)?
```

2. I see zram holds addtional page refcount until EH would complete
the compression, which causes cma allocation latency(easily observed
by page_pinner)

3. some benchmark tools for zram uses direct IO so the long latency
would be not good for them. Yeah, the direct IO is not real workload
at this moment since swap doesn't use it but why not if we could fix
it easily?

Then, why does we select 5-10 range for the sleep?
Just rely on the benchmark, it showed the best result in direct-io
case compared to 0-5, 10-20.

It also showed the 20% EH efficiency regression(iops from 2708 to 2144,
nr_compressed / cpu_tick eh_comp_thread ran) on ACT test but it was
just 0.007% regression from system wide(122 tick / 1673670 tick)

[1] https://www.kernel.org/doc/html/latest/timers/timers-howto.html
Bug: 227623648
Bug: 221039817
Bug: 226436023
Signed-off-by: Minchan Kim <minchan@google.com>
Change-Id: I02d05a29cf1ac0ef66048afe8e719b580ce81102