vh: sched: implement vendor hook to log unfrozen tasks
Freezing of user space tasks may be failed due to process being
in uninterruptible sleep state. Currently, the logging of unfrozen
tasks is disabled by default. Although it can be enabled for
debugging via an adb command, it's usually hard to reproduce the
issue. So add a hook to log unfrozen tasks.

Bug: 240091483
Test: enable the flag via vendor hook and see the logs showing
Change-Id: I67335cc4123344574c6360ccc5ed778893f08a6e
Signed-off-by: Darren Hsu <darrenhsu@google.com>